Extracted each type to its own file
