+++ src/App/Enrollment.cs
@@ -18,13 +18,4 @@
             Grade = grade;
         }
     }
-
-    public enum Grade
-    {
-        A = 0,
-        B = 1,
-        C = 2,
-        D = 3,
-        F = 4
-    }
 }
