+++ src/App/Grade.cs
@@ -1,0 +1,11 @@
+﻿namespace App
+{
+    public enum Grade
+    {
+        A = 0,
+        B = 1,
+        C = 2,
+        D = 3,
+        F = 4
+    }
+}
