+++ src/App/IBus.cs
@@ -1,0 +1,7 @@
+﻿namespace App
+{
+    public interface IBus
+    {
+        void Send(string message);
+    }
+}
